arg: remove ref to ctors in doc
We use arg_t to deduce parameters that *eventually* result in a call to
a ctor, but the actual matching may be against a function like a
factory method. The unit test itself matches against methods in a
handler.

This is also one fewer important word in a place where there are
already too many.