arg_t: fix inverted smf filter
Previously, the arg tests pass but the final integration test was
failing. It turns out the smf filter has been inverted the whole time.

The test was written with the filter inverted, so it was just invoking
the copy ctor. After inverting the filter, the individual tests had to
be modified to correctly convert to the deduced parameter type, not the
final resolved type, since that is the point of the smf filter in the
first place.