Do not explicitly disable resolving aggregates.
Aggregates aren't a special case we need to pick out. We *shouldn't*
resolve a pure aggregate because it would only have elementary types,
and resolving those just returns the default value. It still works,
though, and it even does what you would expect, but this is just not
particularly useful.