.clang-format: make dink headers come first
The variable I was using to store the file list for the
interface lib had the wrong name, so the lib was just empty, failing
silently. Now the name is corrected, so headers are again part of the
project. They missed a few .clang-format changes along the way, but now
they are up to date.