dispatcher_t: cap unresolvables with an assert
Previously, it would just run off the end and say the leaf base class
was undefined. This is true, but not helpful.

Now it at least localizes to a specific specialization with an assert
and a comment.

This isn't the slam-dunk I would hope because neither clang nor cl say
WHY the requirement wasn't met, only that it wasn't.