strong_type_t: allow multiple tags
This way, we can associate a factory's parameterization with its
product, even if that product has a different parameterization.

The motivating example is string_set_t from axiom. It is
not parameterized, but its factory has multiple parameters. The only
way to associate a factory with a string set is for the string set to
have at least as many parameters as the factory.

Previously, strong_type_t accepted only one tag, but that tag could be
a tuple. Now, the tuple is unnecessary.