Add rough support for binding const vs mutable.
This tracks the first call to bind a value. If it is bound const, it
asserts when attempting to retrive it mutable or when trying to rebind.

This is very, very basic, just a proof of concept. Eventually this
should be checked at compile time. For now, though, we can bind consts
and they are, in one sense, protected.