dispatcher: list one reason why resolve may fail
This is a trivial documentation update to describe a scenario that comes
up occasionally that is difficult to diagnose if you don't know why, but
trivial once the possibility is seen.

The reason is a base type defines static construct(), but a leaf does
not, but now this commit message is starting to duplicate the content,
so see the comments added in the commit.