upgrade has_static_construct_method
This now works on templates and construct in bases is no longer a false
positive.

Previously, this detected the presence of static construct() by taking
the address. This only works in a few cases. Now it just tries to call
it in an unevaluated context, which is more canonical and should cover
any case.