dispatcher: reduce compilation error output
This revision puts a bool in both leaves of dispatcher_t to signal
whether dispatch was successful or not, which is static_asserted outside
of the nested dispatchers. This fails the same conditions as before,
but instead of printing a force-multiplied trace through dispatcher_t's
type recursion, it lets that clear out and asserts after.