resolver: rough in scopes
This revision was done a little bit backwards. I have refactored shared_resolver_t to use nested scopes. All *current*
tests pass. None of the new stuff is tested because I wasn't sure this would work this way in the first place. Now that
it appears to, I want to get this into revision control before returning to the new tests.