memory: rename cache_line_size to memory
Originally, I was going to split cache line size and page size into separate functors, but with all of the platform-specific overhead, it is more straightforward to have a memory metrics type that has methods for both.

This revision just changes the file name, but none of the contents, so git tracks it correctly.