memory: refactor in terms of append-only heap allocator
The final object allocator for a di scope needs to be able to allocate for arbitrarily-sized types. Since the paged allocator can only allocate up to page_size at max, a fallback, large-object allocator is necessary. The allocator will compose over both the small-object paged allocator and the large-object heap allocator. Since the paged allocator is append-only, making the heap allocator append-only simplifies the composed allocator. Now that we have an append-only heap allocator, we can implement pages in terms of *that*, which removes a bit of bookkeeping pages were doing.

Now everything is implemented in terms of the append-only heap allocator. This removes the need for owned buffers and the array allocator. The distinctions between the page factory, page allocator, and page buffer source are gone, so I removed the page allocator and buffer source in favor of the page factory, which now works directly in terms of the heap allocator.