tuple: reimplement unique to use key projection
Previously, I was going to use keys_t to extract the keys in a tuple, then work on the keys next to the elements. Instead, the keys can be extracted from head and accumulated. I toyed with using a projection metafunction to specify what to extract, but since this is coupled in the problem domain, it's ok to couple here, too.