bindings: remove tuple machinery
This was all primiarly to support reducing the effective set of bindings by removing overrides and to decrease compile times by using an index map.

There aren't enough overridden bindings in a typical application (at least in examples from other languages where di containers are popular), and we can't use a compile-time map in the way that is necessary until c++26 constexprs typeid. None of this is used now.