config: Add flatten().
This flattens elements and nested tuples into a single flat tuple.

The code starte off in scope/config/binding.hpp, since it will eventually flatten bindings, but it because topical enough for its own module.

Separately, config will not be per scope, but per injector, so it should not go in scope/. It will be a large enough component to get a top level spot, so it now goes in config/.