memory: cast_allocation is a function again
It certainly was not a cpo, but I can't even justify making this a global function object because substituting it isn't useful.
It's only failure condition is if the ctor throws, and that is better simulated in a test by using a type that throws from its ctor, rather than substituting the cast impl.