memory: s/page_size/os_page_size_provider
The arena allocator was originally going to be called "arena_t" and the pool of arenas "pooled_arena_allocator_t". That name is a bit wordy. Now, individual arenas will be called "page_t" and the collection of pages will be called "paged_allocator_t". That would make the name "page_size_t" here confusing, so this revision clarifies that it is the os page size provider.