repo: refactor intrusive list deleters
These need to be tested individually, but this splits the delete_and_deallocate deleter into one that walks the list, a variadic composite, and a dedicated destroy node. The old node deleter has been reimplemented in terms of these and renamed allocated_node_deleter_t, since it applies the allocation deleter as well as the destroy node deleter.