style: west const and pointer alignment
I added a setting to .clang-format to always place const on the left.
This inexplicably gave references and pointers an extra space. After
some googling, I found out the culprit was derived pointer alignment,
so I also disabled that.

I ran all files through the new formatter settings, and this is the
result.