refactor(invoker): order ctor variant first
In the provider, the ctor version is first because that is what will be
used most frequently. In invoker, the factory implementation was first
because it is more generic. Here, the order has been reversed in invoker
so it now has the ctor implementation first, to mirror the order in
provider.