refactor(provider): bash accessors
I always wondered why the accessors were so different from the creators.
It's because they aren't providers. They're scopes. These should be
implemented at the scope level.

After some design work, these are now removed. ExternalReference will
be promoted to an Instance scope. All of the others can be implemented
in terms of the current scopes using a factory function.