fix(scope): remove Requested from cached param
Including the requested type in the Meyers parameterization meant there
was one instance per requested type per provider, not one per provider.
This is now 1:1 with provider, but the types currently must be the
same. Requesting a pointer or shared_ptr will cause a compiler error,
so we assert for now.