feat(container): add transparent tagging
This revision allows users to optionally specify an arbitrary tag on a
container. This can be used to distinguish between two otherwise
identical container types. This is important because caches are keyed by
container and provider, and two containers with the same type use the
same cache. By specifying a tag, the caches can be separated.

It also includes a macro to generate containers with unique tags
because there isn't a way to do it per instance otherwise.