feat(relegation): relegation now returns copies
Singleton relegation used to provide fresh instances directly from the
provider every time a value was requested instead of a reference. Now
it provider is not invoked and a copy of the live singleton reference
value is made instead.

This makes singleton behave the same as instance now when asking for
values.