docs(arity): trial doxygen changes
First, the docblock style changes from from //! to /*!, or we'd have to
use //! on every single documented line, which seems excessive. Now, the
style is to use /* for multiline comments, including docblocks. // or
//! is still used for single lines.

Second, the order of the contents of the file-level docblock are
reordered so the brief is before the copyright. Otherwise, the \brief
ends up being the copyright.